Generate Tide prompt configuration
The Tide propmpt plugin installed for Fish shell requires configuration,
which is stored in Fish variables (see ~/.config/fish/fish_variables).

After a fresh install these variables are not set, therefore the prompt
will break and print errors.

This changes executes the "tide configure --auto" command with the
desired options as CLI arguments, which will generate the config variables.
Benefits:
  - Since the options are parsed, when it is changed by a new Tide version,
    then the "tide configure" command will fail and raise the problem.
  - This method makes sure that the prompt configuration is synchronized
    across machines.

This change also enables displaing the "user@hostname" information. By default
it's shown in the right prompt, but this change moves it to the left prompt.
Unfortunatelt it cannot be set by CLI argument of the configure tool.